1、支持在程序启动时传递-d参数来启用P2P传输模式，启用后数据不再全部交给服务端处理；2、支持通过TCP通信来获取每个节点的在线状态，在转发逻辑中假如在线状态判断，目的地地址状态离线的数据包将不会被转发；3、优化代码结构，各个模块将尽量减少耦合以方便维护。

+++ tapLanSecurity/tapLanSecurity.cpp
@@ -1,7 +1,7 @@
 #include "tapLanSecurity.hpp"
 
-int encryptDataErrCnt = 0;
-int decryptDataErrCnt = 0;
+uint64_t encryptDataErrCnt = 0;
+uint64_t decryptDataErrCnt = 0;
 
 bool tapLanEncryptDataWithAes(uint8_t* data, size_t& dataLen, const TapLanKey& key) {
     bool ret = false;
